Change controllers for auth, change schema for qustion

+++ src/main/java/jenya/gogacpypy/Controllers/AdminController.java
@@ -1,22 +1,31 @@
 package jenya.gogacpypy.Controllers;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jenya.gogacpypy.Utils.JWTProvider;
 import jenya.gogacpypy.model.Admin;
 import jenya.gogacpypy.repository.AdminRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
 @CrossOrigin
+@RequiredArgsConstructor
 public class AdminController {
     @Autowired
     private AdminRepository AdminRepository;
+    private final JWTProvider jwtProvider;
+    private final int levelAccess = 2;
+    private final ObjectMapper mapper;
 
     @GetMapping("/admins")
-    public List<Admin> view_admins(@RequestHeader("Authorization") String token) {
-        return AdminRepository.findAll();
+    public String view_admins(@RequestHeader("Authorization") String token) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(AdminRepository.findAll());
     }
 
     @PostMapping("/add_admin")
@@ -24,6 +33,9 @@ public class AdminController {
                             BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         AdminRepository.save(admin);
         return "{\"res\":\"Success\"}";
@@ -34,6 +46,9 @@ public class AdminController {
                             BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         AdminRepository.deleteById(id);
         return "{\"res\":\"Success\"}";
