+++ src/main/java/jenya/gogacpypy/Controllers/GroupController.java
@@ -1,9 +1,13 @@
 package jenya.gogacpypy.Controllers;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jenya.gogacpypy.Utils.JWTProvider;
 import jenya.gogacpypy.model.Group;
 import jenya.gogacpypy.model.GroupsToQuestions;
 import jenya.gogacpypy.repository.GroupRepository;
 import jenya.gogacpypy.repository.GroupsToQuestionsRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -13,25 +17,38 @@ import java.util.Optional;
 
 @RestController
 @CrossOrigin
+@RequiredArgsConstructor
 public class GroupController {
     @Autowired
     private GroupRepository GroupRepository;
     @Autowired
     private GroupsToQuestionsRepository GroupsToQuestionsRepository;
+    private final JWTProvider jwtProvider;
+    private final int levelAccess = 1;
+    private final ObjectMapper mapper;
 
     @GetMapping("/groups")
-    public List<Group> view_groups() {
-        return GroupRepository.findAll();
+    public String view_groups(@RequestHeader("Authorization") String token) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(GroupRepository.findAll());
     }
 
     @PostMapping("/group_by_id")
-    public Optional<Group> group_by_id(@RequestHeader("Authorization") String token,@RequestBody long id) {
-        return GroupRepository.findById(id);
+    public String group_by_id(@RequestHeader("Authorization") String token,@RequestBody long id) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(GroupRepository.findById(id));
     }
 
     @PostMapping("/groups_by_collection_id")
-    public List<Group> groups_by_collection_id(@RequestHeader("Authorization") String token,@RequestBody long id) {
-        return GroupRepository.findByCollectionId(id);
+    public String groups_by_collection_id(@RequestHeader("Authorization") String token,@RequestBody long id) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(GroupRepository.findByCollectionId(id));
     }
 
     @PostMapping("/add_group")
@@ -39,6 +56,9 @@ public class GroupController {
                                  BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         GroupRepository.save(group);
         return "{\"res\":\"Success\"}";
@@ -50,6 +70,9 @@ public class GroupController {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
         }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
         GroupsToQuestionsRepository.save(groupsToQuestions);
         return "{\"res\":\"Success\"}";
     }
@@ -60,6 +83,9 @@ public class GroupController {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
         }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
         GroupRepository.deleteById(id);
         return "{\"res\":\"Success\"}";
     }
@@ -69,6 +95,9 @@ public class GroupController {
                                                BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         List<GroupsToQuestions> a = GroupsToQuestionsRepository.findByGroupIdAndQuestionId(gtq.getGroupId(),gtq.getQuestionId());
         for (GroupsToQuestions i:
