+++ src/main/java/jenya/gogacpypy/Controllers/TestController.java
@@ -1,9 +1,13 @@
 package jenya.gogacpypy.Controllers;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jenya.gogacpypy.Utils.JWTProvider;
 import jenya.gogacpypy.model.Test;
 import jenya.gogacpypy.model.TestsToGroups;
 import jenya.gogacpypy.repository.TestRepository;
 import jenya.gogacpypy.repository.TestsToGroupsRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -13,30 +17,46 @@ import java.util.Optional;
 
 @RestController
 @CrossOrigin
+@RequiredArgsConstructor
 public class TestController {
     @Autowired
     private TestRepository TestRepository;
     @Autowired
     private TestsToGroupsRepository TestsToGroupsRepository;
+    private final JWTProvider jwtProvider;
+    private final int levelAccess = 1;
+    private final ObjectMapper mapper;
 
     @GetMapping("/tests")
-    public List<Test> view_groups(@RequestHeader("Authorization") String token) {
-        return TestRepository.findAll();
+    public String view_groups(@RequestHeader("Authorization") String token) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(TestRepository.findAll());
     }
 
     @PostMapping("/test_by_id")
-    public Optional<Test> test_by_id(@RequestHeader("Authorization") String token,@RequestBody long id) {
-        return TestRepository.findById(id);
+    public String test_by_id(@RequestHeader("Authorization") String token,@RequestBody long id) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(TestRepository.findById(id));
     }
 
     @PostMapping("/tests_to_groups")
-    public List<TestsToGroups> view_tests_to_groups(@RequestHeader("Authorization") String token,@RequestBody Test test) {
-        return TestsToGroupsRepository.findByTest(test);
+    public String view_tests_to_groups(@RequestHeader("Authorization") String token,@RequestBody Test test) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(TestsToGroupsRepository.findByTest(test));
     }
 
     @PostMapping("/tests_to_groups_by_test_id")
-    public List<TestsToGroups> view_tests_to_groups_by_test_id(@RequestHeader("Authorization") String token,@RequestBody long id) {
-        return TestsToGroupsRepository.findByTestId(id);
+    public String view_tests_to_groups_by_test_id(@RequestHeader("Authorization") String token,@RequestBody long id) throws JsonProcessingException {
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
+        return mapper.writeValueAsString(TestsToGroupsRepository.findByTestId(id));
     }
 
     @PostMapping("/add_test")
@@ -44,6 +64,9 @@ public class TestController {
                             BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         TestRepository.save(test);
         return "{\"res\":\"Success\"}";
@@ -55,6 +78,9 @@ public class TestController {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
         }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
         TestsToGroupsRepository.save(testsToGroups);
         return "{\"res\":\"Success\"}";
     }
@@ -65,6 +91,9 @@ public class TestController {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
         }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
+        }
         TestRepository.deleteById(id);
         return "{\"res\":\"Success\"}";
     }
@@ -74,6 +103,9 @@ public class TestController {
                                         BindingResult result) {
         if (result.hasErrors()) {
             return "{\"res\":\"Have an error\"}";
+        }
+        if (!jwtProvider.checkAccess(token,levelAccess)) {
+            return "{\"res\":\"Access denied\"}";
         }
         TestsToGroupsRepository.deleteById(id);
         return "{\"res\":\"Success\"}";
